Move GsUtil back under IoUtil to cut down on unrelated changes.
